costumizer-ui/pyproject.toml
[tool.poetry]
name = "costumizer-ui"
version = "0.1.0"
description = ""
authors = ["NimajnebEC <nimajnebec@users.noreply.github.com>"]
readme = "../README.md"
packages = [{ include = "costumizer", from = "src/backend" }]

[tool.poetry.dependencies]
python = "^3.11"
mysql-connector-python = "^8.0.32"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
requests = "^2.28.2"
flask = "^2.2.3"
pyjwt = "^2.6.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dev = "scripts:dev"
